Added white piece translation and rotation
Now when the current player is using white, the pieces
are rendered on the bottom row without altering the board state.
This is achieved using translations and rotations to the context of
the canvas. This feature was also included in the drawPiece function to
correctly rotate the drawn pieces after the initial rotations and translations.