Added some check logic
Now it should be that kings that are in check are the only pieces that can move.
It should also know correctly make sure that the king doesnt make moves that will put them
in check. Now all that is needed for implementation is checkmate, draws, castleing, pins, and testing.