Refactored some code and fixed castling
Castling has returned and should be working. Also refactored the isInCheck
method  to just check if the moved piece blocks the king piece if it is in check.
The evaluateGameState method will handle checks, checkmates, and draws. These are the final
parts that need to be implemented along with having a check and updating the board accordingly for
en passant.