Removed check/checkmate bugs and began implementing pawn promotion and draws
There were some edge cases with the doesMoveTakeCheckingPiece and doesMoveBlockCheckingPiece
that resulted in the king not being able to move. Resolved these issues. Pawn promotion basic is now implemented.
This requires a new ui component to be rendered to allow for a piece to change types. Also the whole board needs to be
re-rendered because the screen darkening is a rectangle drawn on top of all the other elements. So I will need to
clear the screen after the selection is made and then re-draw the elements. So far with draws, only fifty moves rule is implemented.
Stalemate, Insufficient Material, and Threefold Repitition are still needed to be implemented.